Remove ptifalls with xdg-basedir support
The old code would happily create $HOME/.xmonad even when the
$XDG_CONFIG_HOME/xmonad directory exists, and is used as source.
This then overrode the config path on the next run, which made XMonad
work once, then break for no apparent reason.
When $XDG_CONFIG_HOME/xmonad exists, it should be safe to asume, the
user wants to use xdg dirs, and they should be prefered when XMonad has
to create directories either way.

The cache dir could result in the same situation, when XMONAD_CONFIG_DIR
and XMONAD_DATA_DIR are set, but XMONAD_CACHE_DIR isn't.
Here it didn't break quite as badly (becaus the env is prefered) but
it's still a directory that probably shouldn't be created, so falling
back to the XMONAD_DATA_DIR (as writeable directory) should prove more
robust in the general case.